refactor: combine RTL and LTR stylesheets (#22526)
## **Description**

Note: I think we need someone with lots of knowledge and experience in
how the app is _supposed_ to look in RTL view to approve this PR.

Pro: Makes language switching faster and simpler while also simplifies
build process. Simplifies the build process and makes it a little bit
faster, too! We now only need one CSS file.

Con: the single `index.css` file is now larger.

This updates the build process to use `gulp-postcss` for our CSS post
processing, and as a byproduct of that I switched it to use `postcss`'s
`autoprefixer` plugin instead of the `gulp-autoprefixer` plugin we use
now. The two plugins do the exact same things, but since we now use
`postcss` it makes sense to use its plugin ecosystem as well.

We only process the CSS once-ish (`postcss` applies 2 plugins, so it's
still almost the same amount of overhead), so it should be slightly
faster to build (I didn't benchmark this though).

### How did it work before?

We generated two separate stylesheets, then switched between them when
the user requires an RTL language.

### How does it work now?

The generated CSS bundle now makes use of our HTML tag's `dir` attribute
(as the good html-lord, Tim Berners-Lee, intended) to switch between
styles (`dir="ltr"` or `dir="rtl"`).

### But Why?

In an effort to speed up our build process (by rewriting it) I'm
modernizing it by using more and performant tools and methods, like
`postcss`. Both build processes will coexist, at least for a while, so
some of the changes necessary for the new build system need to be
applied to the old one as well (or I'd have to complicate the build
further... which I don't want to do!).

Additionally, the new build system parses our HTML files to find all
relevant build assets... because the `rtl` version of our CSS file was
present in our HTML files it would try to parse it... but since it
doesn't exist yet, it would fail.

## **Manual testing steps**

1. Go to the Settings -> General page
2. Change the Language setting to a RTL language
3. Browse around to make sure things aren't broken.