[MMI] Refactored the way we are showing the user the deep custodian link. (#22825)
## **Description**

It appears that the problem we're facing is a result of a code refactor
that took place 6 months ago. During this refactor, the logic that sets
the `custodyId` was removed and set to null. This change has created a
potential race condition where the event handler that adds the
`custodyId` is called after the logic to show the deep link. While the
problem is not significant for GK8, it could potentially affect other
custodians.

Therefore, we have updated the way we call the `show custodian deep
link` function to align with the latest signatureController changes.
This will allow us to obtain the `custodian_transactionId` more
efficiently.

## **Related issues**

Fixes:

## **Manual testing steps**

1. Import some accounts and open up the background inspector, network
tab
2. Create / propose a signed message

## **Screenshots/Recordings**

<!-- If applicable, add screenshots and/or recordings to visualize the
before and after of your change. -->

### **Before**


![image](https://github.com/MetaMask/metamask-extension/assets/1182864/ebe0ffc9-ef9d-4d6a-a4f4-a88c5558dc19)

### **After**

![Screenshot 2024-02-06 at 11 54
04](https://github.com/MetaMask/metamask-extension/assets/1182864/83e6f452-c1af-4c37-b11d-c0e3a11c103e)
![Screenshot 2024-02-06 at 11 53
52](https://github.com/MetaMask/metamask-extension/assets/1182864/b46e0828-e517-4a0a-b686-215cfcb2d371)
![Screenshot 2024-02-06 at 11 53
47](https://github.com/MetaMask/metamask-extension/assets/1182864/94e7afd6-466a-4c3a-a452-e599da4a5b64)


## **Pre-merge author checklist**

- [X] I’ve followed [MetaMask Coding
Standards](https://github.com/MetaMask/metamask-extension/blob/develop/.github/guidelines/CODING_GUIDELINES.md).
- [X] I've clearly explained what problem this PR is solving and how it
is solved.
- [X] I've linked related issues
- [X] I've included manual testing steps
- [X] I've included screenshots/recordings if applicable
- [X] I’ve included tests if applicable
- [X] I’ve documented my code using [JSDoc](https://jsdoc.app/) format
if applicable
- [X] I’ve applied the right labels on the PR (see [labeling
guidelines](https://github.com/MetaMask/metamask-extension/blob/develop/.github/guidelines/LABELING_GUIDELINES.md)).
Not required for external contributors.
- [X] I’ve properly set the pull request status:
  - [X] In case it's not yet "ready for review", I've set it to "draft".
- [x] In case it's "ready for review", I've changed it from "draft" to
"non-draft".

## **Pre-merge reviewer checklist**

- [X] I've manually tested the PR (e.g. pull and build branch, run the
app, test code being changed).
- [X] I confirm that this PR addresses all acceptance criteria described
in the ticket it closes and includes the necessary testing evidence such
as recordings and or screenshots.