Add networkClientId to DetectTokensController.detectNewTokens() (#22898)
## **Description**

Currently networkClientId token based polling in the
`DetectTokensController` does not make the `getBalancesInSingleCall`
contract call with the provider for the networkClientId. This is because
networkClientId is not being passed from `_executePoll` to
`getBalancesInSingleCall`.

This PR adds an optional networkClientId param to `detectNewTokens`
which is used to:
* make `getBalancesInSingleCall` calls with the correct provider
* override the chainId argument if also passed


## **Related issues**

Fixes: https://github.com/MetaMask/MetaMask-planning/issues/2097

## **Manual testing steps**

Nothing should change as this does not affect the current single
globally selected network flow and the multichain polling for this has
not been integrated yet.

## **Screenshots/Recordings**

<!-- If applicable, add screenshots and/or recordings to visualize the
before and after of your change. -->

### **Before**

<!-- [screenshots/recordings] -->

### **After**

<!-- [screenshots/recordings] -->

## **Pre-merge author checklist**

- [x] I’ve followed [MetaMask Coding
Standards](https://github.com/MetaMask/metamask-extension/blob/develop/.github/guidelines/CODING_GUIDELINES.md).
- [x] I've clearly explained what problem this PR is solving and how it
is solved.
- [x] I've linked related issues
- [ ] I've included manual testing steps
- [ ] I've included screenshots/recordings if applicable
- [x] I’ve included tests if applicable
- [ ] I’ve documented my code using [JSDoc](https://jsdoc.app/) format
if applicable
- [x] I’ve applied the right labels on the PR (see [labeling
guidelines](https://github.com/MetaMask/metamask-extension/blob/develop/.github/guidelines/LABELING_GUIDELINES.md)).
Not required for external contributors.
- [x] I’ve properly set the pull request status:
  - [x] In case it's not yet "ready for review", I've set it to "draft".
- [ ] In case it's "ready for review", I've changed it from "draft" to
"non-draft".

## **Pre-merge reviewer checklist**

- [ ] I've manually tested the PR (e.g. pull and build branch, run the
app, test code being changed).
- [ ] I confirm that this PR addresses all acceptance criteria described
in the ticket it closes and includes the necessary testing evidence such
as recordings and or screenshots.