Adds 'yarn master-sync' command to automatically generate a PR-ready master-sync branch (#22052)
## **Description**

Adds a yarn script that can be run from any branch to generate a
master-sync branch. This always has to be done following a merge to
master.

## **Related issues**

Fixes: #

## **Manual testing steps**

On this branch, run `yarn master-sync`. A local master-sync branch will
either be updated or created, and it will be mergable with develop with
only changes to the changelog and the version in the package.json file.

## **Screenshots/Recordings**



https://github.com/MetaMask/metamask-extension/assets/7499938/afbf9f5f-783e-4081-99ba-6eac4ecd37b4





## **Pre-merge author checklist**

- [x] I’ve followed [MetaMask Coding
Standards](https://github.com/MetaMask/metamask-extension/blob/develop/.github/guidelines/CODING_GUIDELINES.md).
- [x] I've clearly explained what problem this PR is solving and how it
is solved.
- [ ] I've linked related issues
- [x] I've included manual testing steps
- [x] I've included screenshots/recordings if applicable
- [ ] I’ve included tests if applicable
- [ ] I’ve documented my code using [JSDoc](https://jsdoc.app/) format
if applicable
- [ ] I’ve applied the right labels on the PR (see [labeling
guidelines](https://github.com/MetaMask/metamask-extension/blob/develop/.github/guidelines/LABELING_GUIDELINES.md)).
Not required for external contributors.
- [ ] I’ve properly set the pull request status:
  - [ ] In case it's not yet "ready for review", I've set it to "draft".
- [ ] In case it's "ready for review", I've changed it from "draft" to
"non-draft".

## **Pre-merge reviewer checklist**

- [ ] I've manually tested the PR (e.g. pull and build branch, run the
app, test code being changed).
- [ ] I confirm that this PR addresses all acceptance criteria described
in the ticket it closes and includes the necessary testing evidence such
as recordings and or screenshots.