gtest: fix several ppom testcases due to description change (#23052)
It seems that the description response for the malicious mocked requests
has changed in a recent update.
This PR aims to fix the e2e tests.

This also reveals the strong need for using a stable version of the cdn,
by:
- Either using the cdn test version of ppom (stable), like we used to do
--> that was changed due to some issue Blockaid had on their end
- Or mocking the cdn responses and keep them always the same --> to
investigate if this is doable, since the responses that we get are in a
"weird format" (see example below)
![Screenshot from 2024-02-19
19-23-25](https://github.com/MetaMask/metamask-extension/assets/54408225/54589804-2c9f-40c7-8f67-2ff800f0b0be)

Another scenario to explore for some of these specs, would be to see if
we can inject the files directly in the ppomDB. In the same way we
initialize the wallet with a certain state, we could initialize the
wallet with the ppomDB prefilled in storage.

This PR however, aims to unblock ci in the fastes possible way

Fixes: https://github.com/MetaMask/metamask-extension/issues/23053

1. Check ci or run them locally

![Screenshot from 2024-02-19
19-04-44](https://github.com/MetaMask/metamask-extension/assets/54408225/248ae58c-fb40-4531-8028-a3e218ed44b3)

<!-- [screenshots/recordings] -->

- [x] I’ve followed [MetaMask Coding
Standards](https://github.com/MetaMask/metamask-extension/blob/develop/.github/guidelines/CODING_GUIDELINES.md).
- [x] I've clearly explained what problem this PR is solving and how it
is solved.
- [x] I've linked related issues
- [x] I've included manual testing steps
- [x] I've included screenshots/recordings if applicable
- [x] I’ve included tests if applicable
- [x] I’ve documented my code using [JSDoc](https://jsdoc.app/) format
if applicable
- [x] I’ve applied the right labels on the PR (see [labeling
guidelines](https://github.com/MetaMask/metamask-extension/blob/develop/.github/guidelines/LABELING_GUIDELINES.md)).
Not required for external contributors.
- [x] I’ve properly set the pull request status:
  - [ ] In case it's not yet "ready for review", I've set it to "draft".
- [ ] In case it's "ready for review", I've changed it from "draft" to
"non-draft".

- [ ] I've manually tested the PR (e.g. pull and build branch, run the
app, test code being changed).
- [ ] I confirm that this PR addresses all acceptance criteria described
in the ticket it closes and includes the necessary testing evidence such
as recordings and or screenshots.